DefaultLeshanCertificateVerifier: Add new certificate verifier helper
This commit just moves inner class implementation from
CaliforniumEndpointsManager to own class.
