+++ leshan-client-cf/src/main/java/org/eclipse/leshan/client/californium/CaliforniumEndpointsManager.java
@@ -134,41 +134,9 @@ public class CaliforniumEndpointsManager implements EndpointsManager {
                 newBuilder.setIdentity(serverInfo.privateKey, new Certificate[] { serverInfo.clientCertificate });
 
                 // set X509 verifier
-                final Certificate expectedServerCertificate = serverInfo.serverCertificate;
-                newBuilder.setCertificateVerifier(new CertificateVerifier() {
-
-                    @Override
-                    public void verifyCertificate(CertificateMessage message, DTLSSession session)
-                            throws HandshakeException {
-                        // As specify in the LWM2M spec 1.0, we only support "domain-issued certificate" usage
-                        // Defined in : https://tools.ietf.org/html/rfc6698#section-2.1.1 (3 -- Certificate usage 3)
-
-                        // Get server certificate from certificate message
-                        if (message.getCertificateChain().getCertificates().size() == 0) {
-                            AlertMessage alert = new AlertMessage(AlertLevel.FATAL, AlertDescription.BAD_CERTIFICATE,
-                                    session.getPeer());
-                            throw new HandshakeException(
-                                    "Certificate chain could not be validated : server cert chain is empty", alert);
-                        }
-                        Certificate receivedServerCertificate = message.getCertificateChain().getCertificates().get(0);
-
-                        // Validate certificate
-                        if (!expectedServerCertificate.equals(receivedServerCertificate)) {
-                            AlertMessage alert = new AlertMessage(AlertLevel.FATAL, AlertDescription.BAD_CERTIFICATE,
-                                    session.getPeer());
-                            throw new HandshakeException(
-                                    "Certificate chain could not be validated: server certificate does not match expected one ('domain-issue certificate' usage)",
-                                    alert);
-                        }
-                    }
-
-                    @Override
-                    public X509Certificate[] getAcceptedIssuers() {
-                        return null;
-                    }
-                });
+                newBuilder.setCertificateVerifier(new DefaultLeshanCertificateVerifier(serverInfo.serverCertificate));
                 serverIdentity = Identity.x509(serverInfo.getAddress(), EndpointContextUtil
-                        .extractCN(((X509Certificate) expectedServerCertificate).getSubjectX500Principal().getName()));
+                        .extractCN(((X509Certificate) serverInfo.serverCertificate).getSubjectX500Principal().getName()));
                 filterCipherSuites(newBuilder,
                         dtlsConfigbuilder.getIncompleteConfig().getSupportedCipherSuites(), false, true);
             } else {
